data-transformation-svc/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiohttp
pdfplumber
openai
python-dotenv